remix-ai-studio/text_to_music_generator/requirements.txt
transformers==4.31.0
torch==2.0.1
pygame==2.5.1
numpy==1.24.3
python-dotenv==1.0.0
